backend/requirements.txt
Django~=4.2.11
Faker~=21.0.1
djangorestframework~=3.14.0
django-filter~=23.5